add tests for complete registration routes

+++ src/main/kotlin/com/admin_bot/features/registration/model/EmailVerifier.kt
@@ -1,7 +1,10 @@
 package com.admin_bot.features.registration.model
 
+import com.admin_bot.common.ExpiredException
+import com.admin_bot.common.ResponseText
 import com.admin_bot.config.ServerConfig
 import com.admin_bot.features.registration.data.OtpData
+import io.ktor.http.*
 import kotlinx.datetime.Clock
 
 /**
@@ -16,17 +19,23 @@ class EmailVerifier(
     suspend fun generateAndSendOtp(botId: Long, email: String) {
         val otp = generateOtp()
         val expiresAt = Clock.System.now().plus(serverConfig.otpLifetime)
-        otpStorage.saveOtpData(botId, OtpData(otp, expiresAt))
+        otpStorage.saveOtpData(botId, OtpData(otp, email, expiresAt))
         verificationEmailSender.sendMailWithOtp(email, expiresAt, otp)
     }
 
-    suspend fun verifyOtp(botId: Long, otp: String):Boolean{
+    /**
+     * Returns confirmed email if otp was verified successfully and null in the other case
+     */
+    suspend fun verifyOtp(botId: Long, otp: String): String? {
         val otpData = otpStorage.getOtpData(botId)
-        if(otpData!=null&&otpData.otpValue==otp){
+        if (otpData != null && otpData.otpValue == otp) {
+            if (otpData.expiresAt < Clock.System.now()) {
+                throw ExpiredException(HttpStatusCode.BadRequest, ResponseText.otpWasExpired)
+            }
             otpStorage.deleteOtpData(botId)
-            return true
+            return otpData.email
         }
-        return false
+        return null
     }
